+++ src/greentest/test__os.py
@@ -1,18 +1,10 @@
 import sys
 import _six as six
 from os import pipe
+import gevent
 from gevent import os
 from greentest import TestCase, main
 from gevent import Greenlet, joinall
-try:
-    import fcntl
-except ImportError:
-    fcntl = None
-
-try:
-    import errno
-except ImportError:
-    errno = None
 
 
 class TestOS_tp(TestCase):
@@ -86,6 +78,30 @@ if hasattr(os, 'make_nonblocking'):
         def write(self, *args):
             return os.nb_write(*args)
 
+
+if hasattr(os, 'fork_and_watch'):
+
+    class TestForkAndWatch(TestCase):
+
+        __timeout__ = 5
+
+        def test_waitpid_all(self):
+            # Cover this specific case.
+            pid = os.fork_and_watch()
+            if pid:
+                os.waitpid(-1, 0)
+                # Can't assert on what the pid actually was,
+                # our testrunner may have spawned multiple children.
+                os._reap_children(0) # make the leakchecker happy
+            else:
+                gevent.sleep(2)
+                os._exit(0)
+
+        def test_waitpid_wrong_neg(self):
+            self.assertRaises(OSError, os.waitpid, -2, 0)
+
+        def test_waitpid_wrong_pos(self):
+            self.assertRaises(OSError, os.waitpid, 1, 0)
 
 if __name__ == '__main__':
     main()